updated to use consolidated k8-types
