+++ src/auth/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fluvio-auth"
-version = "0.2.0"
+version = "0.2.1"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
 edition = "2018"
 license = "Apache-2.0"
@@ -14,7 +14,7 @@ path = "src/lib.rs"
 
 [dependencies]
 async-trait = "0.1.41"
-fluvio-controlplane-metadata = { version = "0.3.0", path = "../controlplane-metadata" }
+fluvio-controlplane-metadata = { version = "0.4.0", path = "../controlplane-metadata" }
 dataplane = { version = "0.1.0", path = "../dataplane-protocol", package = "fluvio-dataplane-protocol" }
 fluvio-future = { version = "0.1.12", features = ["net", "openssl_tls"] }
 fluvio-protocol = { version = "0.2.0" }
