+++ src/cli/Cargo.toml
@@ -45,14 +45,15 @@ tempdir = "0.3.7"
 # Fluvio dependencies
 
 k8-config = { version = "1.3.0" }
-k8-client = { version = "4.0.0" }
+k8-client = { version = "5.0.0" }
 fluvio-future = { version = "0.1.8", features = ["fs", "io", "subscriber"] }
 fluvio = { version = "0.3.0", path = "../client", default-features = false }
 fluvio-cluster = { version = "0.6.0", path = "../cluster", default-features = false, features = ["cli", "platform"] }
 fluvio-package-index = { version = "0.2.0", path = "../package-index" }
 fluvio-extension-consumer = { version = "0.1.2", path = "../extension-consumer" }
 fluvio-extension-common = { version = "0.1.0", path = "../extension-common", features = ["target"]}
-fluvio-controlplane-metadata = { version = "0.3.0", path = "../controlplane-metadata", features = ["use_serde", "k8"] }
+fluvio-controlplane-metadata = { version = "0.4.0", path = "../controlplane-metadata", features = ["use_serde", "k8"] }
+k8-types = { version = "0.1.0", features = ["core"]}
 
 [dev-dependencies]
 fluvio-future = { version = "0.1.0", features = ["fixture"] }
