+++ src/client/Cargo.toml
@@ -36,7 +36,7 @@ semver = "0.11.0"
 # Fluvio dependencies
 fluvio-future = { version = "0.1.10", features = ["task", "native2_tls"] }
 fluvio-types = { version = "0.2.0", path = "../types" }
-fluvio-sc-schema = { version = "0.3.1", path = "../sc-schema", default-features = false }
+fluvio-sc-schema = { version = "0.4.0", path = "../sc-schema", default-features = false }
 fluvio-spu-schema = { version = "0.2.0", path = "../spu-schema" }
 fluvio-socket = { version = "0.4.0", features = ["native_tls"] }
 fluvio-protocol = { version = "0.2.0" }
