+++ src/cluster/Cargo.toml
@@ -50,11 +50,12 @@ fluvio-helm = "0.3.0"
 fluvio-future = { version = "0.1.13" }
 fluvio-runner-local = { version = "0.2.0", path = "../extension-runner-local", optional = true }
 fluvio-extension-common = { version = "0.1.0", path = "../extension-common", optional = true }
-fluvio-controlplane-metadata = { version = "0.3.0", path = "../controlplane-metadata", features = ["k8"] }
+fluvio-controlplane-metadata = { version = "0.4.0", path = "../controlplane-metadata", features = ["k8"] }
 flv-util = "0.5.2"
 k8-config = { version = "1.3.0" }
-k8-client = { version = "4.0.0" }
-k8-metadata-client = { version = "2.0.0" }
+k8-client = { version = "5.0.0" }
+k8-metadata-client = { version = "3.0.0" }
+k8-types = { version = "0.1.0", features = ["app"]}
 
 [dev-dependencies]
 fluvio-future = { version = "0.1.0", features = ["task"] }
