+++ src/controlplane-metadata/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "fluvio-controlplane-metadata"
 edition = "2018"
-version = "0.3.1"
+version = "0.4.0"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
 description = "Fluvio metadata"
 repository = "https://github.com/infinyon/fluvio"
@@ -25,7 +25,7 @@ async-trait = "0.1.21"
 fluvio-future = { version = "0.1.0" }
 flv-util = { version = "0.5.0" }
 fluvio-types = { version = "0.2.0", path = "../types" }
-fluvio-stream-model = { path = "../stream-model", version = "0.3.0" }
+fluvio-stream-model = { path = "../stream-model", version = "0.4.0" }
 fluvio-protocol = { version = "0.2.0" }
 dataplane = { version = "0.1.0", path = "../dataplane-protocol", package = "fluvio-dataplane-protocol" }
