+++ src/controlplane-metadata/src/lib.rs
@@ -11,7 +11,7 @@ pub mod store {
 }
 
 #[cfg(feature = "k8")]
-pub use fluvio_stream_model::k8;
+pub use fluvio_stream_model::k8_types;
 
 pub mod extended {
