+++ src/controlplane-metadata/src/partition/k8.rs
@@ -1,4 +1,4 @@
-use crate::k8::metadata::{Crd, GROUP, V1, CrdNames, Spec, Status, DefaultHeader};
+use crate::k8_types::{Crd, GROUP, V1, CrdNames, Spec, Status, DefaultHeader};
 
 use super::PartitionStatus;
 use super::PartitionSpec;
