+++ src/controlplane-metadata/src/spg/k8/mod.rs
@@ -8,7 +8,7 @@ mod spec;
 pub use self::spec::*;
 
 mod ext {
-    use crate::k8::metadata::Status as K8Status;
+    use crate::k8_types::Status as K8Status;
     use crate::spg::SpuGroupStatus;
 
     /// implement k8 status for spu group status because they are same
