+++ src/controlplane/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fluvio-controlplane"
-version = "0.3.0"
+version = "0.4.0"
 edition = "2018"
 license = "Apache-2.0"
 description = "Fluvio control plane API"
@@ -17,6 +17,6 @@ tracing = "0.1.19"
 
 # Fluvio dependencies
 fluvio-types = { path = "../types", version = "0.2.0" }
-fluvio-controlplane-metadata = { path = "../controlplane-metadata", version = "0.3.0" }
+fluvio-controlplane-metadata = { path = "../controlplane-metadata", version = "0.4.0" }
 fluvio-protocol = { version = "0.2.0" }
 dataplane = { version = "0.1.0", path = "../dataplane-protocol", package = "fluvio-dataplane-protocol" }
