+++ src/extension-consumer/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fluvio-extension-consumer"
-version = "0.1.2"
+version = "0.1.3"
 edition = "2018"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
 description = "Fluvio consumer extension"
@@ -33,5 +33,5 @@ fluvio-future = { version = "0.1.8", features = ["fs", "io"] }
 fluvio = { version = "0.3.0", path = "../client", default-features = false }
 fluvio-types = { version = "0.2.0" , path = "../types" }
 fluvio-extension-common = { version = "0.1.0", path = "../extension-common", features = ["target"] }
-fluvio-controlplane-metadata = { version = "0.3.0", path = "../controlplane-metadata", features = ["use_serde"] }
-fluvio-sc-schema = { version = "0.3.0", path = "../sc-schema", features = ["use_serde"] }
+fluvio-controlplane-metadata = { version = "0.4.0", path = "../controlplane-metadata", features = ["use_serde"] }
+fluvio-sc-schema = { version = "0.4.0", path = "../sc-schema", features = ["use_serde"] }
