+++ src/sc-schema/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fluvio-sc-schema"
-version = "0.3.1"
+version = "0.4.0"
 edition = "2018"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
 description = "Fluvio API for SC"
@@ -23,6 +23,6 @@ static_assertions = "1.1.0"
 
 # Fluvio dependencies
 fluvio-types = { version = "0.2.0", path = "../types" }
-fluvio-controlplane-metadata = { version = "0.3.0", default-features = false, path = "../controlplane-metadata" }
+fluvio-controlplane-metadata = { version = "0.4.0", default-features = false, path = "../controlplane-metadata" }
 fluvio-protocol = { version = "0.2.0" }
 dataplane = { version = "0.1.0", path = "../dataplane-protocol", package = "fluvio-dataplane-protocol" }
