+++ src/sc/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "fluvio-sc"
 edition = "2018"
-version = "0.4.1"
+version = "0.4.2"
 authors = ["fluvio.io"]
 description = "Fluvio Stream Controller"
 repository = "https://github.com/infinyon/fluvio"
@@ -45,13 +45,14 @@ once_cell = "1.5"
 fluvio-auth = { version = "0.2.0", path = "../auth" }
 fluvio-future = { version = "0.1.12", features = ["subscriber", "openssl_tls"] }
 fluvio-types = { version = "0.2.0", path = "../types", features = ["events"] }
-fluvio-sc-schema = { version = "0.3.0", path = "../sc-schema" }
-fluvio-stream-model = { path = "../stream-model", version = "0.3.0" }
-fluvio-controlplane = { version = "0.3.0", path = "../controlplane" }
-fluvio-controlplane-metadata = { version = "0.3.0", features = ["k8", "serde"], path = "../controlplane-metadata" }
-fluvio-stream-dispatcher = { version = "0.3.0", path = "../stream-dispatcher" }
-k8-client = { version = "4.0.0", optional = true }
-k8-metadata-client = { version = "2.0.0" }
+fluvio-sc-schema = { version = "0.4.0", path = "../sc-schema" }
+fluvio-stream-model = { path = "../stream-model", version = "0.4.0" }
+fluvio-controlplane = { version = "0.4.0", path = "../controlplane" }
+fluvio-controlplane-metadata = { version = "0.4.0", features = ["k8", "serde"], path = "../controlplane-metadata" }
+fluvio-stream-dispatcher = { version = "0.4.0", path = "../stream-dispatcher" }
+k8-client = { version = "5.0.0", optional = true }
+k8-metadata-client = { version = "3.0.0" }
+k8-types = { version = "0.1.0", features = ["app"]}
 fluvio-protocol = { version = "0.2.0" }
 dataplane = { version = "0.1.2", path = "../dataplane-protocol", package = "fluvio-dataplane-protocol" }
 fluvio-socket = { version = "0.4.2" }
