+++ src/spu/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "fluvio-spu"
 edition = "2018"
-version = "0.3.2"
+version = "0.3.3"
 authors = ["fluvio.io"]
 description = "Fluvio Stream Processing Unit"
 repository = "https://github.com/infinyon/fluvio"
@@ -41,8 +41,8 @@ async-io = "1.3.1"
 # Fluvio dependencies
 fluvio-types = { version = "0.2.0", path = "../types" }
 fluvio-storage = { version = "0.2.0", path = "../storage" }
-fluvio-controlplane = { version = "0.3.0", path = "../controlplane" }
-fluvio-controlplane-metadata = { version = "0.3.0", path = "../controlplane-metadata" }
+fluvio-controlplane = { version = "0.4.0", path = "../controlplane" }
+fluvio-controlplane-metadata = { version = "0.4.0", path = "../controlplane-metadata" }
 fluvio-spu-schema = { version = "0.2.0", path = "../spu-schema" }
 fluvio-protocol = { version = "0.2.0" }
 dataplane = { version = "0.1.0", path = "../dataplane-protocol", package = "fluvio-dataplane-protocol" }
