+++ src/stream-dispatcher/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "fluvio-stream-dispatcher"
 edition = "2018"
-version = "0.3.2"
+version = "0.4.0"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
 description = "Fluvio Event Stream access"
 repository = "https://github.com/infinyon/fluvio"
@@ -27,8 +27,8 @@ once_cell = "1.5"
 
 # Fluvio dependencies
 fluvio-types = { path = "../types", version = "0.2.0" }
-fluvio-stream-model = { features = ["k8"], version = "0.3.0", path = "../stream-model" }
-k8-metadata-client = { version = "2.0.0" }
+fluvio-stream-model = { features = ["k8"], version = "0.4.0", path = "../stream-model" }
+k8-metadata-client = { version = "3.0.0" }
 fluvio-future = { version = "0.1.10" }
 
 [dev-dependencies]
