+++ src/stream-model/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "fluvio-stream-model"
 edition = "2018"
-version = "0.3.1"
+version = "0.4.0"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
 description = "Fluvio Event Stream Model"
 repository = "https://github.com/infinyon/fluvio"
@@ -13,7 +13,7 @@ path = "src/lib.rs"
 
 [features]
 use_serde = ["serde"]
-k8 = ["use_serde", "k8-obj-app"]
+k8 = ["use_serde", "k8-types"]
 
 [dependencies]
 tracing = "0.1.19"
@@ -23,7 +23,7 @@ event-listener = "2.5.1"
 once_cell = "1.5"
 
 # Fluvio dependencies
-k8-obj-app = { version = "2.0.0", optional = true }
+k8-types = { version = "0.1.0", optional = true }
 
 
 [dev-dependencies]
