+++ src/stream-model/src/store/k8.rs
@@ -10,7 +10,7 @@ use std::cmp::PartialEq;
 
 use tracing::error;
 
-use crate::k8::app::core::metadata::{Spec as K8Spec, Status as K8Status, ObjectMeta, K8Obj};
+use crate::k8_types::{Spec as K8Spec, Status as K8Status, ObjectMeta, K8Obj};
 use crate::store::{MetadataStoreObject};
 use crate::core::{Spec, MetadataItem, MetadataContext};
